src: add internal erofs writer code
This introduces experimental code for writing erofs images ourselves,
instead of using the external mkcomposefs CLI.

It's currently disabled by default.  You can test it by setting the
`COMPOSEFS_FORMAT=new` environment variable.

This currently produces a different output than the output of
mkcomposefs, which is why it's gated behind an environment variable.
The plan is to add a compatibility mode to our internal writer code so
that it produces as similar of an output as possible and then switch
over to using it once we are convinced that it's equivalent.  Then the
`COMPOSEFS_FORMAT=` variable will disable this compatibility mode.

There's also a `COMPOSEFS_DUMP_EROFS=1` environment variable (which
works with both `mkcomposefs` and our internal code) which will dump the
erofs layout for diffing.  There's also a standalone `erofs-debug`
binary that will do the same.

Additionally, this introduces two new files in docs:
 - a detailed description of the parts of erofs that we use
 - a document which attempts to describe the decisions made in creating
   an erofs composefs image (in terms of which order the files are in,
   etc).

The main idea here is to start a serious effort towards standardizing
the composefs label we want to start adding to container images: it
should be possible to define what will be in that label by way of
documentation instead of saying "run this software and use the output".
These two new documents, taken together with the existing "oci.md" form
a rough (and still incomplete) outline for that.

Many thanks to Gao Xiang <xiang@kernel.org> for helping clarify many
points about the erofs file format for the documentation.

Closes #56

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>